requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.1
    # via streamlit
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   pytest
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
narwhals==2.9.0
    # via altair
numpy==2.3.4
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==25.0
    # via
    #   altair
    #   pytest
    #   streamlit
pandas==2.3.3
    # via
    #   lab-03 (pyproject.toml)
    #   streamlit
pillow==11.3.0
    # via streamlit
pluggy==1.6.0
    # via pytest
protobuf==6.33.0
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via lab-03 (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.50.0
    # via lab-03 (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
